유저 로그인 연동 및 및 예외처리

+++ src/main/java/com/dali/dali/domain/community/service/CommunityService.java
@@ -10,6 +10,7 @@ import com.dali.dali.domain.park.Park;
 import com.dali.dali.domain.park.ParkRepository;
 import com.dali.dali.domain.users.entity.User;
 import com.dali.dali.domain.users.repository.UserRepository;
+import com.dali.dali.global.exception.UnauthorizedException;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -19,6 +20,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.Optional;
 
@@ -31,12 +33,15 @@ public class CommunityService {
     private final UserRepository userRepository;
 
     @Transactional
-    public Community createPost(CommunityDto communityDto) {
+    public Community createPost(CommunityDto communityDto, Principal principal) {
+
+        if (principal == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
 
         Park park = parkRepository.findById(communityDto.getPark_id())
                 .orElseThrow(() -> new EntityNotFoundException(communityDto.getPark_id() + ": 공원을 찾을 수 없습니다."));
 
-        // 로그인 구현 후 현재 로그인한 사용자로 검증
         User user = userRepository.findById(communityDto.getUser_id())
                 .orElseThrow(() -> new EntityNotFoundException(communityDto.getUser_id() + ": 유저를 찾을 수 없습니다."));
 
@@ -55,13 +60,24 @@ public class CommunityService {
         return communityRepository.save(community);
     }
 
-    public Community getPost(Long id) {
+    public Community getPost(Long id, Principal principal) {
+
+        if (principal == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
         return communityRepository.findById(id).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "게시글이 존재하지 않습니다.")
         );
     }
 
-    public Page<Community> getPosts(Pageable pageable, AMPM ampm, Time time, Gender gender, String park_name) {
+    public Page<Community> getPosts(Pageable pageable, AMPM ampm,
+                                    Time time, Gender gender, String park_name, Principal principal) {
+
+        if (principal == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
         if (ampm == null && time == null && gender == null && park_name == null) {
             return communityRepository.findAll(pageable);
         } else if (gender != null) {
@@ -77,7 +93,7 @@ public class CommunityService {
     }
 
     @Transactional
-    public Community updatePost(Long id, CommunityDto communityDto) {
+    public Community updatePost(Long id, CommunityDto communityDto, Principal principal) {
         Optional<Community> existPost = communityRepository.findById(id);
         if (existPost.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정할 게시글이 존재하지 않습니다.");
@@ -86,8 +102,20 @@ public class CommunityService {
         Park park = parkRepository.findById(communityDto.getPark_id())
                 .orElseThrow(() -> new EntityNotFoundException(communityDto.getPark_id() + ": 공원을 찾을 수 없습니다."));
 
+        if (principal == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
+        String email = principal.getName();
+        Optional<User> optionalUser = userRepository.findByEmail(email);
+        User loginUser = optionalUser.get();
+        User communityUser = existPost.get().getUser();
+
+        if (!loginUser.getEmail().equals(communityUser.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "작성자만 게시글 수정이 가능합니다.");
+        }
+
         Community community = Community.builder()
-                .id(existPost.get().getId())
                 .title(communityDto.getTitle())
                 .content(communityDto.getContent())
                 .gender(communityDto.getGender())
@@ -103,10 +131,24 @@ public class CommunityService {
     }
 
     @Transactional
-    public Community deletePost(Long id) {
+    public Community deletePost(Long id, Principal principal) {
         Community community = communityRepository.findById(id).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "삭제할 게시글이 존재하지 않습니다.")
         );
+
+        if (principal == null) {
+            throw new UnauthorizedException("로그인이 필요합니다.");
+        }
+
+        String email = principal.getName();
+        Optional<User> optionalUser = userRepository.findByEmail(email);
+        User loginUser = optionalUser.get();
+        User communityUser = community.getUser();
+
+        if (!loginUser.getEmail().equals(communityUser.getEmail())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "작성자만 게시글 삭제가 가능합니다.");
+        }
+
         communityRepository.deleteById(id);
         return community;
     }
